+++ Views/SettingsWindow.axaml.cs
@@ -9,8 +9,9 @@ using FluentIcons.Common;
 using System;
 using System.Diagnostics;
 using System.Linq;
+using DMXVideoPlayer.Services;
 
-namespace DMXVideoPlayer
+namespace DMXVideoPlayer.Views
 {
     public partial class SettingsWindow : Window
     {